+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The next release will require at least [Go 1.25].
 - Add `const Version` in `go.opentelemetry.io/contrib/propagators/autoprop`. (#8488)
 - Add `const Version` in `go.opentelemetry.io/contrib/processors/minsev`. (#8590)
 - Add `const Version` in `go.opentelemetry.io/contrib/exporters/autoexport`. (#8611)
+- Host resource detector in `go.opentelemetry.io/contrib/otelconf` now includes `os.` attributes. (#8578)
 
 ### Fixed